[src & egs] Replace WSJ0-mix Dataset (#97)
