+++ asteroid/data/__init__.py
@@ -2,3 +2,4 @@ from .wham_dataset import WhamDataset
 from .whamr_dataset import WhamRDataset
 from .dns_dataset import DNSDataset
 from .librimix_dataset import LibriMix
+from .wsj0_mix import Wsj0mixDataset
